moved shortenDate fucntion to formatDate.js
