+++ services/utils/dateFormat.js
@@ -1,4 +1,9 @@
-export default function dateFormat({ format = "long", date, locale = "en-GB", UTCLocal = false }) {
+export default function dateFormat({
+  format = "long",
+  date,
+  locale = "en-GB",
+  UTCLocal = false,
+}) {
   let dateTimeStyle;
   let formatLocale = locale;
   let formatDate = date;
@@ -21,9 +26,28 @@ export default function dateFormat({ format = "long", date, locale = "en-GB", UT
   if (UTCLocal) {
     dateTimeStyle = {
       ...dateTimeStyle,
-      timeZone: 'UTC'
-    }
+      timeZone: "UTC",
+    };
   }
 
-  return new Intl.DateTimeFormat(formatLocale, dateTimeStyle).format(new Date(formatDate));
+  return new Intl.DateTimeFormat(formatLocale, dateTimeStyle).format(
+    new Date(formatDate),
+  );
+}
+
+export function shortenDate({ date, formatStyle }) {
+  const milestoneDate = new Date(date);
+  const day = milestoneDate.getDate();
+  const month = milestoneDate.getMonth() + 1;
+  const year = milestoneDate.getFullYear();
+  switch (formatStyle) {
+    case "dd/mm/yyyy":
+      return `${day}/${month}/${year}`;
+    case "month/year":
+      return `${month}/${year}`;
+    case "year":
+      return `${year}`;
+    default:
+      return `${day}/${month}/${year}`;
+  }
 }
